convert old UMC paths from data/campaigns/ to data/add-ons/ in a mainline-safe way
My first step is to draw up a list of mainline campaigns to compare path directories against.

As for the conversion itself, I want wmllint to be able to accurately handle a possible case where a line contains both a UMC and a mainline path - even though I don't know of any actual example in the wild. So:

cleanup_main: This will be used to convert mainline paths back to normal, as we shall see.
while: As long as there is any instance of "data/campaigns/<dir>", we will keep replacing.
  dc : This regex object uses the same regex as the while statement, except that the campaign's directory name has been separated out as a group and extended. (It would get the job done with just '\w+', since all mainline campaigns use only letters and underscores, but I've added characters that are used in some UMC names so that the full directory name gets printed to stdout.)
  if: group(1) is checked against the list of mainline campaigns. If it's not there, the line is rebuilt substituting 'data/add-ons/' where 'data/campaigns/' was, and a report printed to stdout.
  else: The alternative means that it is a mainline path. However, we need the 'while' test to go forward, so in this case, a meaningless string is substituted for 'data/campaigns/'. The line is flagged by setting cleanup_main to True.
if cleanup_main: Now we can change mainline paths back to "data/campaigns", a simple replace operation. This means that mainline campaign files will not be touched, since the net effect is to leave the line unchanged.