check for textdomain paths that don't include /translations, and binary paths referring to 'external' data
The in_textdomain and in_binary_path code can also be used to check path directories.

First, the convention that [textdomain] uses "/translations" is strongly established, and I can't think of a legitimate reason for an add-on not to be following it. So if it doesn't, let's print an alert to stdout.

As for binary paths, there is the continuing use of special directories to hold menu images. Although these are non-standard, the names that took hold are "/public" and variations of "/external*". So as a crude test, we can look for those strings, and print an error. (A more sophisticated test would look for binary paths outside the campaign #ifdef.) It does not catch the worst case of all - when all binaries are outside the campaign define, not just a set-aside directory.